chore: add helper functions for batteries
